requirements.txt
Django==4.1
djangorestframework==3.13.1
django-cors-headers==3.13.0
flake8==5.0.4
black==22.8.0
django-environ==0.9.0
haversine==2.6.0